made changes to better reflect wireframe
