+++ munchmap-app/app/layout.tsx
@@ -3,7 +3,7 @@ import { Inter } from "next/font/google";
 import "./globals.css";
 import "@mantine/core/styles.css";
 import { createTheme, MantineProvider } from "@mantine/core";
-import { HeaderNoGrid } from "./components/Header";
+import { Header } from '../components/Header/Header';
 
 const theme = createTheme({
   /** Put your mantine theme override here */
@@ -12,7 +12,7 @@ const theme = createTheme({
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "Munch Map",
   description: "Generated by create next app",
 };
 
@@ -25,8 +25,8 @@ export default function RootLayout({
     <html lang="en">
       <body className={inter.className} style={{backgroundColor: "#b4d9b0"}}>
         <MantineProvider theme={theme}>
-          {/* <Head>{children}</Head> */}
-          <HeaderNoGrid>{children}</HeaderNoGrid>
+      
+          <Header>{children}</Header>
         </MantineProvider>
       </body>
     </html>
