+++ munchmap-app/components/Header/Header.module.css
@@ -7,6 +7,7 @@
   padding-right: var(--mantine-spacing-md);
   background-color: #008195;
   color: white;
+
 }
 
 .header > div > div {
@@ -22,7 +23,7 @@
   text-decoration: none;
   color: light(var(--mantine-color-black), var(--mantine-color-white));
   font-weight: 500;
-  font-size: var(--mantine-font-size-sm);
+  font-size: var(--mantine-font-size-md);
 
   @media (max-width: $mantine-breakpoint-sm) {
     height: rem(42px);
@@ -34,8 +35,7 @@
   }
 
   [data-mantine-color-scheme] &[data-active] {
-    background-color: var(--mantine-color-blue-filled);
-    color: var(--mantine-color-white);
+    color: #FCC419;
   }
 }
